requirements.txt
streamlit
openai
aiohttp
psycopg2-binary
aiomysql